+++ ft_transcendence_db/src/entity/channels_users/channels_users.gateway.ts
@@ -1,6 +1,7 @@
 import { WebSocketGateway, SubscribeMessage, MessageBody, WebSocketServer, ConnectedSocket } from '@nestjs/websockets';
 import { ChannelsUsersService } from './channels_users.service';
 import { Server } from 'socket.io';
+import { SessionService } from '../sessions/session.service';
 
 @WebSocketGateway(3001, {
   cors: {
@@ -13,16 +14,22 @@ export class ChannelsUsersGateway {
   @WebSocketServer()
   server: Server;
 
-  constructor(private readonly channelsUsersService: ChannelsUsersService) {}
+  constructor(private readonly channelsUsersService: ChannelsUsersService, private readonly sessionService: SessionService) {}
 
   /**
    * Gets the users that are in the channel passed as param
    * 
-   * @param body - {channel}
+   * @param body - {channel, sessionCookie}
    * @emits 'updateListUsers' {users, channel}
    */
   @SubscribeMessage('findUsersOfChannel')
-  async findUsersOfChannel(@MessageBody() body) {
+  async findUsersOfChannel(@MessageBody() body)
+  {
+    if (await this.sessionService.getIsSessionExpired(body.sessionCookie))
+    {
+      // TODO redirect to log page
+      return ('not connected');
+    }
     try 
     {
         const res = await this.channelsUsersService.findUsersOfChannel(body.channel.name);
@@ -35,13 +42,18 @@ export class ChannelsUsersGateway {
   /**
    * Modify the relation of a user to a channel to set isBanned = true 
    * 
-   * @param {Object} body - {user, channel}
+   * @param {Object} body - {user, channel, sessionCookie}
    * @returns result of request
    * @emits updateAfterPart {users, channel, user}
    */
   @SubscribeMessage('banUser')
   async ban(@MessageBody() body) 
   {
+    if (await this.sessionService.getIsSessionExpired(body.sessionCookie))
+    {
+      // TODO redirect to log page
+      return ('not connected');
+    }
     const res = await this.channelsUsersService.ban(body.channel, body.user);
     const users = await this.channelsUsersService.findUsersOfChannel(body.channel.name);
     this.server.emit('updateAfterPart', {
@@ -53,13 +65,18 @@ export class ChannelsUsersGateway {
   /**
    * Delete the relation of a user to a channel
    * 
-   * @param body - {user, channel}
+   * @param body - {user, channel, sessionCookie}
    * @returns result of request kick
    * @emits updateAfterPart {users, channel, user}
    */
   @SubscribeMessage('kickUser')
   async kick(@MessageBody() body) 
   {
+    if (await this.sessionService.getIsSessionExpired(body.sessionCookie))
+    {
+      // TODO redirect to log page
+      return ('not connected');
+    }
     const res = await this.channelsUsersService.leave(body.channel, body.user);
     const users = await this.channelsUsersService.findUsersOfChannel(body.channel.name);
     this.server.emit('updateAfterPart', {
@@ -71,13 +88,18 @@ export class ChannelsUsersGateway {
   /**
    * Delete the relation of a user to a channel
    * 
-   * @param body - {user, channel}
+   * @param body - {user, channel, sessionCookie}
    * @returns result of request leave
    * @emits updateAfterPart {users, channel, user}
    */
   @SubscribeMessage('leaveChannel')
   async leaveChannel(@MessageBody() body) 
   {
+    if (await this.sessionService.getIsSessionExpired(body.sessionCookie))
+    {
+      // TODO redirect to log page
+      return ('not connected');
+    }
     const res = await this.channelsUsersService.leave(body.channel, body.user);
     const users = await this.channelsUsersService.findUsersOfChannel(body.channel.name);
     this.server.emit('updateAfterPart', {
@@ -89,13 +111,18 @@ export class ChannelsUsersGateway {
   /**
    * set user of channel to Admin
    * 
-   * @param body - {channel, user}
+   * @param body - {channel, user, sessionCookie}
    * @returns result of request
    * @emits updateListUsers {users, channel}
    */
   @SubscribeMessage('setAdmin')
   async setAdmin(@MessageBody() body) 
   {
+    if (await this.sessionService.getIsSessionExpired(body.sessionCookie))
+    {
+      // TODO redirect to log page
+      return ('not connected');
+    }
     const res = await this.channelsUsersService.setAdmin(body.channel, body.user);
     const users = await this.channelsUsersService.findUsersOfChannel(body.channel.name);
 
@@ -108,13 +135,18 @@ export class ChannelsUsersGateway {
   /**
    * remove admin rights on a user
    * 
-   * @param body - {channel, user}
+   * @param body - {channel, user, sessionCookie}
    * @returns result of request
    * @emits updateListUsers {users, channel}
    */
   @SubscribeMessage('removeAdmin')
   async removeAdmin(@MessageBody() body) 
   {
+    if (await this.sessionService.getIsSessionExpired(body.sessionCookie))
+    {
+      // TODO redirect to log page
+      return ('not connected');
+    }
     const res = await this.channelsUsersService.removeAdmin(body.channel, body.user);
     const users = await this.channelsUsersService.findUsersOfChannel(body.channel.name);
 
@@ -125,7 +157,7 @@ export class ChannelsUsersGateway {
   }
   /**
    * 
-   * @param body {user, userTimeouted, duration_timeout}
+   * @param body {user, userTimeouted, duration_timeout, sessionCookie}
    * @emits timeoutWrongAmount {channel, user}
    * @emits updateListUsers {channel, users}
    * @emits timeoutGoodRequest {channel, user}
@@ -133,6 +165,11 @@ export class ChannelsUsersGateway {
   @SubscribeMessage('timeoutUser')
   async timeoutUser(@MessageBody() body)
   {
+    if (await this.sessionService.getIsSessionExpired(body.sessionCookie))
+    {
+      // TODO redirect to log page
+      return ('not connected');
+    }
     if (body.duration_timeout >= 2592000 || body.duration_timeout < 10) // 30 days and 10 seconds
     {
       this.server.emit('timeoutWrongAmount', {channel: body.channel, user: body.user});
